added tests for texture in AutoRally cost function

+++ MPPI-Generic/kernel_tests/cost_functions/ar_standard_cost_kernel_test.cuh
@@ -16,8 +16,8 @@ void launchCheckCudaArray(std::vector<float4>& result_arr, cudaArray* array, int
 
 __global__ void checkCudaArrayKernel(float4* result_arr, cudaArray* array, int number);
 
-void launchTextureTestKernel(const ARStandardCost& cost, std::vector<float4>& test_results, std::vector<int2>& test_indexes);
+void launchTextureTestKernel(const ARStandardCost& cost, std::vector<float4>& test_results, std::vector<float2>& test_indexes);
 
-__global__ void textureTestKernel(ARStandardCost& cost, float4* test_results, int2* test_indexes, int num_points);
+__global__ void textureTestKernel(ARStandardCost& cost, float4* test_results, float2* test_indexes, int num_points);
 
 #endif //MPPIGENERIC_AR_STANDARD_COST_KERNEL_TEST_CUH
